+++ exporter/opentelemetry-exporter-zipkin/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Change package name to opentelemetry-exporter-zipkin
   ([#953](https://github.com/open-telemetry/opentelemetry-python/pull/953))
+- Add proper length zero padding to hex strings of traceId, spanId, parentId sent on the wire, for compatibility with jaeger-collector
+  ([#908](https://github.com/open-telemetry/opentelemetry-python/pull/908))
 
 ## 0.8b0